adding new doodle classifier example
Dear Diary,

I spent the day recording a set of video tutorials about how to do image classification with a simple "toy" neural network. There are lots of problems with this and the goal is purely educational and to have some fun. I hope to improve this example in the future. It could use an intrface and some animations, the softmax function, and working with a larger dataset and ml5/deeplearn.js.

yours forever,
The Coding Train (choo choo)